Adding comments and clean coding
